requirements.txt
boto3
pytest
pydantic
cloudscale-sdk